Crate pallet_drand

Drand Bridge Pallet

A pallet to bridge to drand’s Quicknet, injecting publicly verifiable randomness into the runtime

Overview

Quicknet chain runs in an ‘unchained’ mode, producing a fresh pulse of randomness every 3s This pallet implements an offchain worker that consumes pulses from quicket and then sends a signed transaction to encode them in the runtime. The runtime uses the optimized arkworks host functions to efficiently verify the pulse.

Run cargo doc --package pallet-drand --open to view this pallet’s documentation.

Re-exports

pub use pallet::*;

Modules

bls12_381

crypto

Based on the above KeyTypeId we need to generate a pallet-specific crypto type wrappers. We can use from supported crypto kinds (sr25519, ed25519 and ecdsa) and augment the types with this pallet-specific identifier.

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

types

utils

verifier

A collection of verifiers

Constants

KEY_TYPE

Defines application identifier for crypto keys of this module.

MAX_PULSES_TO_FETCH

QUICKNET_CHAIN_HASH

the drand quicknet chain hash quicknet uses ‘Tiny’ BLS381, with small 48-byte sigs in G1 and 96-byte pubkeys in G2

Functions

message

construct a message (e.g. signed by drand)