pallet_drand::pallet

Trait Config

pub trait Config: CreateSignedTransaction<Call<Self>> + Config {
    type AuthorityId: AppCrypto<Self::Public, Self::Signature>;
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type Verifier: Verifier;
    type UnsignedPriority: Get<TransactionPriority>;
    type HttpFetchTimeout: Get<u64>;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type AuthorityId: AppCrypto<Self::Public, Self::Signature>

The identifier type for an offchain worker.

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

The overarching runtime event type.

type Verifier: Verifier

something that knows how to verify beacon pulses

type UnsignedPriority: Get<TransactionPriority>

A configuration for base priority of unsigned transactions.

This is exposed so that it can be tuned for particular runtime, when multiple pallets send unsigned transactions.

type HttpFetchTimeout: Get<u64>

The maximum number of milliseconds we are willing to wait for the HTTP request to complete.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.