pallet_drand

Module types

Structs

BeaconConfiguration

A drand chain configuration

BeaconConfigurationPayload

Payload used by to hold the beacon config required to submit a transaction.

BeaconInfoResponse

the expected response body from the drand api endpoint api.drand.sh/{chainId}/info

DrandResponseBody

a pulse from the drand beacon the expected response body from the drand api endpoint api.drand.sh/{chainId}/public/latest

Metadata

metadata for the drand beacon configuration

MetadataInfoResponse

metadata associated with the drand info response

Pulse

A pulse from the drand beacon

PulsesPayload

Payload used by to hold the pulse data required to submit a transaction.

Type Aliases

BoundedHash

an opaque hash type

OpaquePublicKey

Represents an opaque public key used in drand’s quicknet

RoundNumber

the round number to track rounds of the beacon